moved Raiting component to separate folder
